Extract DI tokens to dedicated injector files
